feat(embeds): add embed api, with revoke, get, getall, and create
