+++ lib/api/index.js
@@ -1,6 +1,7 @@
 var api = {
     comment: require("./comment"),
     downtime: require("./downtime"),
+    embed: require("./embed"),
     event: require("./event"),
     graph: require("./graph"),
     host: require("./host"),
